major updates for the change log

+++ platform/core/setting/src/Http/Controllers/SettingController.php
@@ -2,44 +2,55 @@
 
 namespace Botble\Setting\Http\Controllers;
 
-use Assets;
-use BaseHelper;
+use Botble\Base\Exceptions\LicenseIsAlreadyActivatedException;
+use Botble\Base\Facades\Assets;
+use Botble\Base\Facades\BaseHelper;
+use Botble\Base\Facades\EmailHandler;
+use Botble\Base\Facades\PageTitle;
 use Botble\Base\Http\Controllers\BaseController;
 use Botble\Base\Http\Responses\BaseHttpResponse;
 use Botble\Base\Supports\Core;
 use Botble\Base\Supports\Language;
+use Botble\JsValidation\Facades\JsValidator;
+use Botble\Media\Facades\RvMedia;
 use Botble\Media\Repositories\Interfaces\MediaFileInterface;
+use Botble\Media\Repositories\Interfaces\MediaFolderInterface;
+use Botble\Setting\Facades\Setting;
+use Botble\Setting\Http\Requests\EmailSettingRequest;
 use Botble\Setting\Http\Requests\EmailTemplateRequest;
 use Botble\Setting\Http\Requests\LicenseSettingRequest;
 use Botble\Setting\Http\Requests\MediaSettingRequest;
 use Botble\Setting\Http\Requests\ResetEmailTemplateRequest;
 use Botble\Setting\Http\Requests\SendTestEmailRequest;
 use Botble\Setting\Http\Requests\SettingRequest;
-use Botble\Setting\Repositories\Interfaces\SettingInterface;
 use Carbon\Carbon;
-use EmailHandler;
 use Exception;
+use Illuminate\Console\Application;
+use Illuminate\Contracts\View\View;
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\File;
-use RvMedia;
+use Illuminate\Support\ProcessUtils;
 use Throwable;
 
 class SettingController extends BaseController
 {
-    public function __construct(protected SettingInterface $settingRepository)
-    {
-    }
-
     public function getOptions()
     {
-        page_title()->setTitle(trans('core/setting::setting.title'));
+        PageTitle::setTitle(trans('core/setting::setting.title'));
 
-        Assets::addScriptsDirectly('vendor/core/core/setting/js/setting.js')
-            ->addStylesDirectly('vendor/core/core/setting/css/setting.css')
-            ->usingVueJS();
+        Assets::addScripts(['jquery-validation', 'form-validation'])
+            ->addScriptsDirectly([
+                'vendor/core/core/setting/js/setting.js',
+                'vendor/core/core/setting/js/verify-license.js',
+            ])
+            ->addStylesDirectly('vendor/core/core/setting/css/setting.css');
 
-        return view('core/setting::index');
+        Assets::usingVueJS();
+
+        $jsValidation = JsValidator::formRequest(SettingRequest::class);
+
+        return view('core/setting::index', compact('jsValidation'));
     }
 
     public function postEdit(SettingRequest $request, BaseHttpResponse $response)
@@ -58,10 +69,10 @@ class SettingController extends BaseController
             session()->put('site-locale', $locale);
         }
 
-        $isDemoModeEnabled = app()->environment('demo');
+        $isDemoModeEnabled = BaseHelper::hasDemoModeEnabled();
 
         if (! $isDemoModeEnabled) {
-            setting()->set('locale', $locale)->save();
+            setting()->set('locale', $locale);
         }
 
         $adminTheme = $request->input('default_admin_theme');
@@ -70,7 +81,7 @@ class SettingController extends BaseController
         }
 
         if (! $isDemoModeEnabled) {
-            setting()->set('default_admin_theme', $adminTheme)->save();
+            setting()->set('default_admin_theme', $adminTheme);
         }
 
         $adminLocalDirection = $request->input('admin_locale_direction');
@@ -79,7 +90,8 @@ class SettingController extends BaseController
         }
 
         if (! $isDemoModeEnabled) {
-            setting()->set('admin_locale_direction', $adminLocalDirection)->save();
+            setting()->set('admin_locale_direction', $adminLocalDirection);
+            setting()->save();
         }
 
         return $response
@@ -102,15 +114,18 @@ class SettingController extends BaseController
 
     public function getEmailConfig()
     {
-        page_title()->setTitle(trans('core/base::layouts.setting_email'));
+        PageTitle::setTitle(trans('core/base::layouts.setting_email'));
 
         Assets::addScriptsDirectly('vendor/core/core/setting/js/setting.js')
-            ->addStylesDirectly('vendor/core/core/setting/css/setting.css');
+            ->addStylesDirectly('vendor/core/core/setting/css/setting.css')
+            ->addScripts(['jquery-validation', 'form-validation']);
 
-        return view('core/setting::email');
+        $jsValidation = JsValidator::formRequest(EmailSettingRequest::class);
+
+        return view('core/setting::email', compact('jsValidation'));
     }
 
-    public function postEditEmailConfig(SettingRequest $request, BaseHttpResponse $response)
+    public function postEditEmailConfig(EmailSettingRequest $request, BaseHttpResponse $response)
     {
         $this->saveSettings($request->except(['_token']));
 
@@ -121,7 +136,7 @@ class SettingController extends BaseController
 
     public function getEditEmailTemplate(string $type, string $module, string $template)
     {
-        page_title()->setTitle(trans(config($type . '.' . $module . '.email.templates.' . $template . '.title', '')));
+        PageTitle::setTitle(trans(config($type . '.' . $module . '.email.templates.' . $template . '.title', '')));
 
         Assets::addStylesDirectly([
             'vendor/core/core/base/libraries/codemirror/lib/codemirror.css',
@@ -165,19 +180,34 @@ class SettingController extends BaseController
 
     public function postResetToDefault(ResetEmailTemplateRequest $request, BaseHttpResponse $response)
     {
-        $this->settingRepository->deleteBy(['key' => $request->input('email_subject_key')]);
+        Setting::delete([$request->input('email_subject_key')]);
 
         $templatePath = get_setting_email_template_path($request->input('module'), $request->input('template_file'));
 
-        File::delete($templatePath);
+        if (File::exists($templatePath)) {
+            File::delete($templatePath);
+        }
+
+        $shouldBeCleanedDirectories = [
+            File::dirname($templatePath),
+            storage_path('app/email-templates'),
+        ];
+
+        foreach ($shouldBeCleanedDirectories as $shouldBeCleanedDirectory) {
+            if (File::isDirectory($shouldBeCleanedDirectory) && File::isEmptyDirectory($shouldBeCleanedDirectory)) {
+                File::deleteDirectory($shouldBeCleanedDirectory);
+            }
+        }
 
         return $response->setMessage(trans('core/setting::setting.email.reset_success'));
     }
 
     public function postChangeEmailStatus(Request $request, BaseHttpResponse $response)
     {
+        $request->validate(['key' => 'string', 'value' => 'in:0,1']);
+
         setting()
-            ->set($request->input('key'), $request->input('value'))
+            ->set($request->input('key'), $request->boolean('value'))
             ->save();
 
         return $response->setMessage(trans('core/base::notices.update_success_message'));
@@ -202,14 +232,21 @@ class SettingController extends BaseController
         }
     }
 
-    public function getMediaSetting()
+    public function getMediaSetting(MediaFolderInterface $mediaFolderRepository)
     {
-        page_title()->setTitle(trans('core/setting::setting.media.title'));
+        PageTitle::setTitle(trans('core/setting::setting.media.title'));
 
         Assets::addScriptsDirectly('vendor/core/core/setting/js/setting.js')
-            ->addStylesDirectly('vendor/core/core/setting/css/setting.css');
+            ->addStylesDirectly('vendor/core/core/setting/css/setting.css')
+            ->addScripts(['jquery-validation', 'form-validation']);
 
-        return view('core/setting::media');
+        $folderIds = json_decode((string)setting('media_folders_can_add_watermark'), true);
+
+        $folders = $mediaFolderRepository->pluck('name', 'id', ['parent_id' => 0]);
+
+        $jsValidation = JsValidator::formRequest(MediaSettingRequest::class);
+
+        return view('core/setting::media', compact('folders', 'folderIds', 'jsValidation'));
     }
 
     public function postEditMediaSetting(MediaSettingRequest $request, BaseHttpResponse $response)
@@ -221,23 +258,32 @@ class SettingController extends BaseController
             ->setMessage(trans('core/base::notices.update_success_message'));
     }
 
-    public function getVerifyLicense(Core $coreApi, BaseHttpResponse $response)
+    public function getVerifyLicense(Request $request, Core $core, BaseHttpResponse $response)
     {
-        if (! File::exists(storage_path('.license'))) {
-            return $response->setError()->setMessage('Your license is invalid. Please activate your license!');
+        if ($request->expectsJson() && ! $core->checkConnection()) {
+            return response()->json([
+                'message' => __('Your server is not connected to the internet.'),
+            ], 400);
+        }
+
+        $invalidMessage = 'Your license is invalid. Please activate your license!';
+
+        $licenseFilePath = $core->getLicenseFilePath();
+
+        if (! File::exists($licenseFilePath)) {
+            return $response->setError()->setMessage($invalidMessage);
         }
 
         try {
-            $result = $coreApi->verifyLicense(true);
-
-            if (! $result['status']) {
-                return $response->setError()->setMessage($result['message']);
+            if (! $core->verifyLicense(true)) {
+                return $response->setError()->setMessage($invalidMessage);
             }
 
-            $activatedAt = Carbon::createFromTimestamp(filectime($coreApi->getLicenseFilePath()));
+            $activatedAt = Carbon::createFromTimestamp(filectime($core->getLicenseFilePath()));
+            $message = 'Your license is activated.';
         } catch (Throwable $exception) {
             $activatedAt = Carbon::now();
-            $result = ['message' => $exception->getMessage()];
+            $message = $exception->getMessage();
         }
 
         $data = [
@@ -245,10 +291,10 @@ class SettingController extends BaseController
             'licensed_to' => setting('licensed_to'),
         ];
 
-        return $response->setMessage($result['message'])->setData($data);
+        return $response->setMessage($message)->setData($data);
     }
 
-    public function activateLicense(LicenseSettingRequest $request, BaseHttpResponse $response, Core $coreApi)
+    public function activateLicense(LicenseSettingRequest $request, BaseHttpResponse $response, Core $core)
     {
         $buyer = $request->input('buyer');
         if (filter_var($buyer, FILTER_VALIDATE_URL)) {
@@ -257,49 +303,37 @@ class SettingController extends BaseController
 
             return $response
                 ->setError()
-                ->setMessage('Envato username must not a URL. Please try with username "' . $username . '"!');
+                ->setMessage(sprintf('Envato username must not a URL. Please try with username "%s"!', $username));
         }
 
+        $purchasedCode = $request->input('purchase_code');
+
         try {
-            $purchaseCode = $request->input('purchase_code');
+            if (! $core->activateLicense($purchasedCode, $buyer)) {
+                return $response->setError()->setMessage('Your license is invalid.');
+            }
 
-            $result = $coreApi->activateLicense($purchaseCode, $buyer);
+            $data = $this->saveActivatedLicense($core, $buyer);
 
-            $resetLicense = false;
-            if (! $result['status']) {
-                if (str_contains($result['message'], 'License is already active')) {
-                    $coreApi->deactivateLicense($purchaseCode, $buyer);
+            return $response->setMessage('Your license has been activated successfully!')->setData($data);
+        } catch (LicenseIsAlreadyActivatedException) {
+            try {
+                $core->revokeLicense($purchasedCode, $buyer);
 
-                    $result = $coreApi->activateLicense($purchaseCode, $buyer);
-
-                    if (! $result['status']) {
-                        return $response->setError()->setMessage($result['message']);
-                    }
-
-                    $resetLicense = true;
-                } else {
-                    return $response->setError()->setMessage($result['message']);
+                if (! $core->activateLicense($purchasedCode, $buyer)) {
+                    return $response->setError()->setMessage('Your license is invalid.');
                 }
+
+                $data = $this->saveActivatedLicense($core, $buyer);
+
+                return $response
+                    ->setMessage('Your license has been activated successfully and the license on the previous domain has been revoked!')
+                    ->setData($data);
+            } catch (LicenseIsAlreadyActivatedException) {
+                return $response
+                    ->setError()
+                    ->setMessage('Exceeded maximum number of activations, please contact our support to reset your license.');
             }
-
-            setting()
-                ->set(['licensed_to' => $buyer])
-                ->save();
-
-            $activatedAt = Carbon::createFromTimestamp(filectime($coreApi->getLicenseFilePath()));
-
-            $data = [
-                'activated_at' => $activatedAt->format('M d Y'),
-                'licensed_to' => $buyer,
-            ];
-
-            if ($resetLicense) {
-                return $response->setMessage(
-                    $result['message'] . ' Your license on the previous domain has been revoked!'
-                )->setData($data);
-            }
-
-            return $response->setMessage($result['message'])->setData($data);
         } catch (Throwable $exception) {
             return $response
                 ->setError()
@@ -307,43 +341,29 @@ class SettingController extends BaseController
         }
     }
 
-    public function deactivateLicense(BaseHttpResponse $response, Core $coreApi)
+    public function deactivateLicense(BaseHttpResponse $response, Core $core)
     {
         try {
-            $result = $coreApi->deactivateLicense();
+            $core->deactivateLicense();
 
-            if (! $result['status']) {
-                if (str_contains($result['message'], 'Activation not recognized or is inactive')) {
-                    $this->settingRepository->deleteBy(['key' => 'licensed_to']);
+            Setting::delete(['licensed_to']);
 
-                    File::delete(storage_path('.license'));
-
-                    return $response->setMessage('Deactivated license successfully!');
-                }
-
-                return $response->setError()->setMessage($result['message']);
-            }
-
-            $this->settingRepository->deleteBy(['key' => 'licensed_to']);
-
-            return $response->setMessage($result['message']);
+            return $response->setMessage('Deactivated license successfully!');
         } catch (Throwable $exception) {
             return $response->setError()->setMessage($exception->getMessage());
         }
     }
 
-    public function resetLicense(LicenseSettingRequest $request, BaseHttpResponse $response, Core $coreApi)
+    public function resetLicense(LicenseSettingRequest $request, BaseHttpResponse $response, Core $core)
     {
         try {
-            $result = $coreApi->deactivateLicense($request->input('purchase_code'), $request->input('buyer'));
-
-            if (! $result['status']) {
-                return $response->setError()->setMessage($result['message']);
+            if (! $core->revokeLicense($request->input('purchase_code'), $request->input('buyer'))) {
+                return $response->setError()->setMessage('Could not reset your license.');
             }
 
-            $this->settingRepository->deleteBy(['key' => 'licensed_to']);
+            Setting::delete(['licensed_to']);
 
-            return $response->setMessage($result['message']);
+            return $response->setMessage('Your license has been reset successfully.');
         } catch (Throwable $exception) {
             return $response->setError()->setMessage($exception->getMessage());
         }
@@ -441,5 +461,40 @@ class SettingController extends BaseController
         $content = $emailHandler->prepareData($content);
 
         return BaseHelper::clean($content);
+    }
+
+    public function cronjob(): View
+    {
+        PageTitle::setTitle(trans('core/setting::setting.cronjob.name'));
+
+        Assets::addScriptsDirectly('vendor/core/core/setting/js/setting.js');
+
+        $command = sprintf(
+            '* * * * * cd %s && %s >> /dev/null 2>&1',
+            BaseHelper::hasDemoModeEnabled() ? 'path-to-your-project' : ProcessUtils::escapeArgument(base_path()),
+            Application::formatCommandString('schedule:run')
+        );
+
+        $lastRunAt = Setting::get('cronjob_last_run_at');
+
+        if ($lastRunAt) {
+            $lastRunAt = Carbon::parse($lastRunAt);
+        }
+
+        return view('core/setting::cronjob', compact('command', 'lastRunAt'));
+    }
+
+    protected function saveActivatedLicense(Core $core, string $buyer): array
+    {
+        setting()
+            ->set(['licensed_to' => $buyer])
+            ->save();
+
+        $activatedAt = Carbon::createFromTimestamp(filectime($core->getLicenseFilePath()));
+
+        return [
+            'activated_at' => $activatedAt->format('M d Y'),
+            'licensed_to' => $buyer,
+        ];
     }
 }
